+++ packages/tools/pie-icons-webc/vite.config.js
@@ -55,4 +55,14 @@ export default defineConfig({
             },
         },
     },
+    test: {
+        dir: '.',
+        environment: 'jsdom',
+        globals: true,
+        exclude: [
+            '**/test/{accessibility,component,system,visual}/*.spec.{js,ts}',
+            '**/test/mixins/**/*.browser.spec.{js,ts}',
+            '**/node_modules/**/*'
+        ],
+    },
 });