+++ packages/tools/pie-icons/src/__tests__/index.test.js
@@ -1,4 +1,3 @@
-/* eslint-env jest */
 import pieIcons from '../index';
 
 test('has correct properties', () => {