+++ apps/pie-storybook/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.61.2
+
+### Patch Changes
+
+- Updated dependencies [[`30b5ef87c`](https://github.com/justeattakeaway/pie/commit/30b5ef87c81c5943e65fb3d86579fb26fcb1f7d2), [`839e19e4d`](https://github.com/justeattakeaway/pie/commit/839e19e4d926a6386aae988b6e9231c03296de5a)]:
+  - @justeattakeaway/pie-cookie-banner@0.17.0
+  - @justeattakeaway/pie-modal@0.38.4
+
+## 0.61.1
+
+### Patch Changes
+
+- Updated dependencies [[`68828edd7`](https://github.com/justeattakeaway/pie/commit/68828edd730a001acf599a0b6717f43e56d58a18), [`04891f5a0`](https://github.com/justeattakeaway/pie/commit/04891f5a0bd9d880b731db3bf15d87600e2b59d2)]:
+  - @justeattakeaway/pie-tag@0.5.0
+
 ## 0.61.0
 
 ### Minor Changes